app/requirements.txt
flask
pillow
torch
torchvision
requests